app: move ConnMgr to app.internal.connmgr
